+++ type-decls.d.ts
@@ -69,3 +69,9 @@ declare module 'eslint-plugin-cypress' {
   declare const value: ESLint.ESLint.Plugin;
   export = value;
 }
+
+declare module 'eslint-plugin-no-unsanitized' {
+  import type ESLint from 'eslint';
+  declare const value: ESLint.ESLint.Plugin;
+  export = value;
+}